Remove \t shortcut in edition mode
It's not at all uncommon to type a tab escape sequence in a translated
message. The \t shortcut interferes with the \t key sequence by taking you
out of the edition mode and moving the cursor to the Last-Translator header
line.

The \t shortcut still set the translator name in normal mode.